requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
openpyxl>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.0
apify-client>=1.7.0
